Fix missing "Log in" link
There was an issue where the "Log in" link was not showing up under the
"Create Account" button. This was due to styling being applied to all
anchor elements instead of just the anchor elements for the photo
credit. Added a class to the photo credit links to fix this.